mesh: Move atomic.h to sources
This header is not a part of Mesh API.
