+++ nimble/host/mesh/src/atomic.h
